Fixed Drop-off Obstacle Issue
Tested different situations to recreate the dropoff bug with obstacles present. Initially the obstacle would prevent the rover from ever reaching a drop-off state. Created a variable that keeps track of whether or not it needs to use its ultrasound. If the conditions are met such that it needs to drop something off, it will reset drop-off to kill the interrupt. Need further testing to validate fix.