Partial fixes for dropoff
Working on multiple issues. 1) rover getting stuck on the edge of the collection zone because it saw too many cubes. 2) reduced the rover driving speed for debugging. 3) Added a number of output statement to track the state of the rovers controllers and PIDs. 4) Set a minimum obstacle avoidance distance so the cube doesn't cause the rover to enter obstacle avoidance.
t