First cut implementation of site-fidelity.
Currently the robots always return to the last location they found a
resource. If the encounter an obstacle along the way, then they
abandon site-fidelity. This is probably not the behavior we want,
since once recruitment is added, they will encounder many robots while
returning to a cluster (for example). Still a pretty good functional first-cut.