WIP: Updates AddWaypoint Method
Makes it so that it doesn't add a cluster way point if it is already trying
to get to a cluster or if it has site fidelity. Doesn't appear to be working at the moment
it is unclear as to why. Imagine it has to do with someting in the logic controller