Cleans up the code for DropOff
There is was a fair amount of repetitive if statements,
clearing local variables and repeated evaluations of values.
One thing to note is that the spin no longer takes the increment and
divides it by 8 every time, now it just increments and I divided the value
there by 8. No reason to do something in runtime that can be done staticly